+++ spessasynth_core/src/synthesizer/audio_engine/engine_methods/controller_control/controller_change.c
@@ -51,12 +51,12 @@ void ss_channel_set_generator_override(SS_MIDIChannel *ch, SS_GeneratorType gen,
 }
 
 static void ss_channel_update_tuning(SS_MIDIChannel *ch) {
-	ch->channel_tuning_cents =
+	ch->channel_tuning_cents = (int)(
 	ch->custom_controllers[SS_CUSTOM_CTRL_TUNING] + /* RPN channel fine tuning */
 	ch->custom_controllers[SS_CUSTOM_CTRL_TRANSPOSE_FINE] + /* User tuning (transpose) */
 	ch->custom_controllers[SS_CUSTOM_CTRL_MASTER_TUNING] + /* Master tuning, set by sysEx */
 	ch->custom_controllers[SS_CUSTOM_CTRL_TUNING_SEMITONES] *
-	100; /* RPN channel coarse tuning */
+	100.0f); /* RPN channel coarse tuning */
 }
 
 void ss_channel_set_custom_controller(SS_MIDIChannel *ch, SS_CustomController type, float val) {
@@ -65,13 +65,13 @@ void ss_channel_set_custom_controller(SS_MIDIChannel *ch, SS_CustomController ty
 }
 
 void ss_channel_set_tuning(SS_MIDIChannel *ch, float cents) {
-	cents = round(cents);
+	cents = roundf(cents);
 	ss_channel_set_custom_controller(ch, SS_CUSTOM_CTRL_TUNING, cents);
 }
 
 void ss_channel_set_modulation_depth(SS_MIDIChannel *ch, float cents) {
-	cents = round(cents);
-	ss_channel_set_custom_controller(ch, SS_CUSTOM_CTRL_MODULATION_MULTIPLIER, cents / 50.0);
+	cents = roundf(cents);
+	ss_channel_set_custom_controller(ch, SS_CUSTOM_CTRL_MODULATION_MULTIPLIER, cents / 50.0f);
 }
 
 void ss_channel_controller(SS_MIDIChannel *ch, int cc, int val, double time) {
